+++ src/main/java/hu/nye/algterv/transfersystem/model/region/Country.java
@@ -1,0 +1,45 @@
+package hu.nye.algterv.transfersystem.model.regions;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "countries")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Country {
+    
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer countryId;
+
+    private String countryName;
+
+    @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
+    @JoinColumn(name = "region_id", referencedColumnName = "regionId")
+    private GeoRegion regionId;
+
+    private Integer settlementNumber;
+
+    private Integer airportNumber;
+
+    private Integer trainStationNumber;
+
+    private Integer busStationNumber;
+
+    private Integer shipStationNumber;
+}
