+++ src/main/java/hu/nye/algterv/transfersystem/model/region/GeoRegion.java
